+++ Lab1_Encapsulation_Inheritance_polymorphism/Lab1_Encapsulation_Inheritance_polymorphism/Models/Triangle.cs
@@ -149,7 +149,7 @@ namespace Lab1_Encapsulation_Inheritance_polymorphism.Models
 			Console.WriteLine($"Paint {nameof(Triangle)}");
 		}
 
-		public override void Move(int x, int y)
+		public override void Move(double x, double y)
 		{
 			Console.WriteLine($"Move {nameof(Triangle)}");
 		}
