+++ Lab1_Encapsulation_Inheritance_polymorphism/Lab1_Encapsulation_Inheritance_polymorphism/interfaces/IMoveable.cs
@@ -6,6 +6,6 @@ namespace Lab1_Encapsulation_Inheritance_polymorphism.interfaces
 	public interface IMoveable
 	{
 		event Func<Point, Polygon, bool> FigureMoved;
-		void Move(int x, int y);
+		void Move(double x, double y);
 	}
 }
